requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
psycopg2-binary==2.8.4
pytz==2020.1
sqlparse==0.3.1
